# apps/web/src/shared/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fleet Maintenance Decision API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/components/{component_id}/imports/csv": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Import Csv",
        "operationId": "import_csv_api_components__component_id__imports_csv_post",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Component Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CsvImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Import Csv Api Components  Component Id  Imports Csv Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/components/{component_id}/imports": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Import Component",
        "operationId": "import_component_api_components__component_id__imports_post",
        "parameters": [
          {
            "name": "component_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Component Id"
            }
          },
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoryImport"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Import Component Api Components  Component Id  Imports Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/models/registrations": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Model Registration",
        "operationId": "model_registration_api_models_registrations_post",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Model Registration Api Models Registrations Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/jobs/assessment": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Submit Assessment",
        "operationId": "submit_assessment_api_jobs_assessment_post",
        "parameters": [
          {
            "name": "x-demo-user",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo-planner",
              "title": "X-Demo-User"
            }
          },
          {
            "name": "x-demo-role",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "planner",
              "title": "X-Demo-Role"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit Assessment Api Jobs Assessment Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AssessmentRequest": {
        "properties": {
          "import_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Import Id"
          },
          "model_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Model Id"
          },
          "cutoff_cycle": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cutoff Cycle"
          }
        },
        "type": "object",
        "required": [
          "import_id",
          "model_id",
          "cutoff_cycle"
        ],
        "title": "AssessmentRequest"
      },
...
      "CsvImportRequest": {
        "properties": {
          "csv_text": {
            "type": "string",
            "maxLength": 1500000,
            "minLength": 1,
            "title": "Csv Text"
          },
          "source_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Source Version"
          },
          "engine_identity": {
            "type": "string",
            "maxLength": 120,
            "title": "Engine Identity"
          },
          "previous_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Id"
          }
        },
        "type": "object",
        "required": [
          "csv_text",
          "source_version",
          "engine_identity"
        ],
        "title": "CsvImportRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoryImport": {
        "properties": {
          "source_version": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Source Version"
          },
          "engine_identity": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^NASA_CMAPSS:FD001:(train|test):[1-9][0-9]*$",
            "title": "Engine Identity"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/HistoryRow"
            },
            "type": "array",
            "maxItems": 10000,
            "minItems": 1,
            "title": "Rows"
          },
          "previous_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_version",
          "engine_identity",
          "rows"
        ],
        "title": "HistoryImport"
      },
      "HistoryRow": {
        "properties": {
          "cycle": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Cycle"
          },
          "values": {
            "items": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "array",
            "maxItems": 24,
            "minItems": 24,
            "title": "Values"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cycle",
          "values"
        ],
        "title": "HistoryRow"
      },
...
      "Registration": {
        "properties": {
          "version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.-]+$",
            "title": "Version"
          },
          "directory": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Directory"
          }
        },
        "type": "object",
        "required": [
          "version",
          "directory"
        ],
        "title": "Registration"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}